refactor!: use is now generating the shell commands
The use executable is now generating the shell commands directly on the
output, instead of using some tags that need parsing.

Integration of a new shell now requires editing the code itself, but it
gives more flexibility to the code.

Another change is that the clink script is now generating a batch file
that is then called to, avoid empty lines on the output.